requirements.txt
# DGRCL Strategy Dependencies

torch>=2.0.0
torch_geometric>=2.4.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
yfinance>=0.2.0
requests>=2.28.0
lxml>=4.9.0
pytest>=7.0.0

# Phase 1: Portfolio Construction
cvxpy>=1.4.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Live Paper Trading
# NOTE: alpaca-trade-api requires websockets<11, but yfinance requires websockets>=13.
# Both work in practice because alpaca only uses websockets for streaming (unused here).
# Install order: alpaca first, then `pip install "websockets>=13"` to restore yfinance.
alpaca-trade-api>=3.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
